Added typecast to 64 bits or upper bits are lost in mpz_get_uint64() on 32 bit machines
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@859 de4796e2-0b1e-0410-91b4-86bc23549317

+++ utils/gmp_aux.c
@@ -31,7 +31,7 @@ mpz_get_uint64 (mpz_t z)
       ASSERT_ALWAYS (sizeof (unsigned long int) == 4);
       q = mpz_get_ui (z); /* get the low word of z */
       mpz_div_2exp (z, z, 32);
-      q += mpz_get_ui (z) << 32;
+      q += (uint64_t) mpz_get_ui (z) << 32;
     }
   return q;
 }
